+++ packages/yl_player/test/player_controller_test.dart
@@ -1,8 +1,5 @@
-import 'dart:async';
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:yl_player/yl_player.dart';
-import 'package:yl_player_platform_interface/yl_player_platform_interface.dart';
 
 import 'support/fake_player_platform.dart';